+++ src/background.js
@@ -138,6 +138,10 @@ chrome.runtime.onMessage.addListener((request, sender, sendResponse) => {
 		// Updates (overwriting videos) a playlist in Firebase
 		case "overwritePlaylistInfoInDB":
 			updatePlaylistInfoInDB('uploadsPlaylists/' + request.data.key, request.data.val, true).then(sendResponse);
+			break;		
+		// Before v1.0.0 the videos were stored in an array without upload times, so they need to all be refetched
+		case 'updateDBPlaylistToV1.0.0':
+			updateDBPlaylistToV1_0_0('uploadsPlaylists/' + request.data.key).then(sendResponse);
 			break;
 		// Gets an API key depending on user settings
 		case "getAPIKey":
@@ -169,7 +173,7 @@ chrome.runtime.onMessage.addListener((request, sender, sendResponse) => {
 
 // ---------- Firebase ----------
 import { initializeApp } from 'firebase/app';
-import { getDatabase, ref, child, update, get } from 'firebase/database';
+import { getDatabase, ref, child, update, get, remove } from 'firebase/database';
 
 const firebaseConfig = {
 	apiKey: "AIzaSyA6d7Ahi7fMB4Ey8xXM8f9C9Iya97IGs-c",
@@ -192,7 +196,7 @@ async function updatePlaylistInfoInDB(playlistId, playlistInfo, overwriteVideos)
 	} else {
 		console.log("Updating playlistInfo in the database...");
 		// Contains all properties except the videos
-		const playlistInfoWithoutVideos = Object.fromEntries(Object.entries(playlistInfo).filter(([key, value]) => key !== "videos"));
+		const playlistInfoWithoutVideos = Object.fromEntries(Object.entries(playlistInfo).filter(([key, value]) => (key !== "videos")));
 
 		// Upload the 'metadata'
 		update(ref(db, playlistId), playlistInfoWithoutVideos);
@@ -202,6 +206,13 @@ async function updatePlaylistInfoInDB(playlistId, playlistInfo, overwriteVideos)
 	}
 
 	return "PlaylistInfo was sent to database.";
+}
+
+async function updateDBPlaylistToV1_0_0(playlistId) {
+	// Remove all videos from the database
+	remove(ref(db, playlistId + '/videos'));
+
+	return "Videos were removed from the database playlist.";
 }
 
 // Prefers to get cached data instead of sending a request to the database